refactor: split docs graph export helpers
